+++ PersonalBudgetDb/Entities/MonthlyIncome.cs
@@ -2,9 +2,6 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace PersonalBudgetDb.Entities
 {
@@ -18,7 +15,7 @@ namespace PersonalBudgetDb.Entities
         [Required]
         public Guid Id { get; set; }
 
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
 
         public string IdentityUserId { get; set; }
