+++ PersonalBudgetDb/Entities/MonthlyIncomeType.cs
@@ -1,10 +1,6 @@
 ﻿using Microsoft.AspNetCore.Identity;
 using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace PersonalBudgetDb.Entities
 {
@@ -18,7 +14,7 @@ namespace PersonalBudgetDb.Entities
         [MaxLength(30)]
         public string Description { get; set; }
 
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
 
         public string IdentityUserId { get; set; }
