+++ PersonalBudgetDb/Entities/YearlyIncome.cs
@@ -2,9 +2,6 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace PersonalBudgetDb.Entities
 {
@@ -14,13 +11,14 @@ namespace PersonalBudgetDb.Entities
         {
             this.MonthlyIncomes = new HashSet<MonthlyIncome>();
         }
+
         [Key]
         [Required]
         public Guid Id { get; set; }
 
         public int Year { get; set; }
 
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
 
         public string IdentityUserId  { get; set; }
